+++ src/main/java/lermitage/intellij/extra/icons/ExtraIconProvider.java
@@ -785,6 +785,12 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                     .eq("package.json"),
                 ofFile("packagejsonlock", "extra-icons/packagejsonlock.png", "NPM: package-lock.json")
                     .eq("package-lock.json"),
+                ofFile("poetry-poetry-lock", "extra-icons/poetry_lock.svg", "Poetry: poetry.lock")
+                    .eq("poetry.lock"),
+                ofFile("poetry-poetry-toml", "extra-icons/poetry.svg", "Poetry: poetry.toml")
+                    .eq("poetry.toml"),
+                ofFile("poetry-pyproject-toml", "extra-icons/poetry.svg", "Poetry: pyproject.toml")
+                    .eq("pyproject.toml"),
                 ofFile("pdd", "extra-icons/pdd.svg", "Puzzle Driven Development: .pdd")
                     .eq(".pdd"),
                 ofFile("pdd_yml", "extra-icons/pdd.svg", "Puzzle Driven Development: .0pdd.yml")